+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # CHANGELOG
 
+## 0.2.0 (2012-07-02)
+
+* You can now pass a custom set of keys to be used in the `aggregate` call if you want to
+use a different set of positive/negative things than what is setup in the configuration.
+
 ## 0.1.0
 
 * Key-space is now separated by a single `:` instead of `::`
 
-## 0.0.2 
+## 0.0.2
 
 * Fix +total_key+ configuration.